+++ src/facet/helpers/__init__.py
@@ -2,8 +2,9 @@
 Helper processors and utilities.
 """
 
-from .interactive import WaitForConfirmation
+from .interactive import ArtifactOffsetFinder, WaitForConfirmation
 
 __all__ = [
+    'ArtifactOffsetFinder',
     'WaitForConfirmation',
 ]